add stable/custom fip for debug and retrieve
The fip provided with OpenWRT is lacking a few features that are beneficial for offloading data and providing additional debug support. This customized fip adds a few new tools including memory tests, TFTP put, and http get (wget) as well as being slightly more verbose under trouble conditions.

The fip is based on the OpenWRT 23.05.3 UBI configuration for the Linksys e8450, so it should be safe to use when debugging any router running the stable UBI layout up through 23.05.x.